docs: Update notification samples to v1
