+++ samples/v1/receiveNotifications.js
